Consolidate get_provider_fees methods into 1
